dummy player damage added

+++ Unity Project/Assets/Scripts/UI/HealthBar.cs
@@ -5,13 +5,13 @@ public class HealthBar : MonoBehaviour
 {
     public Slider slider;
 
-    public void SetMaxHealth(int health)
+    public void SetMaxHealth(short health)
     {
         slider.maxValue = health;
         slider.value = health;
     }
     
-    public void SetHealth(int health)
+    public void SetHealth(short health)
     {
         slider.value = health;
     }
